+++ guizhanlib-localization/src/main/java/net/guizhanss/guizhanlib/localization/package-info.java
@@ -1,4 +1,4 @@
 /**
- * This package provides localization services.
+ * This package provides single-file localization service.
  */
 package net.guizhanss.guizhanlib.localization;