+++ guizhanlib-minecraft/src/main/java/net/guizhanss/guizhanlib/minecraft/commands/AbstractCommand.java
@@ -27,6 +27,7 @@ import java.util.stream.Collectors;
 @SuppressWarnings("ConstantConditions")
 @Getter
 public abstract class AbstractCommand {
+
     private final Set<SubCommand> subCommands = new HashSet<>();
     private final AbstractCommand parent;
     private final String name;