+++ guizhanlib-minecraft/src/main/java/net/guizhanss/guizhanlib/minecraft/commands/SubCommand.java
@@ -8,6 +8,7 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.function.BiFunction;
 
 public abstract class SubCommand extends AbstractCommand {
+
     protected SubCommand(@Nullable AbstractCommand parent, @Nonnull String name,
                          @Nonnull BiFunction<AbstractCommand, CommandSender, String> description,
                          @Nonnull String usage, @Nonnull SubCommand... subCommands) {