+++ guizhanlib-minecraft/src/main/java/net/guizhanss/guizhanlib/minecraft/helper/attribute/package-info.java
@@ -1,4 +1,0 @@
-/**
- * 该包提供了属性相关的语言助手
- */
-package net.guizhanss.guizhanlib.minecraft.helper.attribute;