+++ guizhanlib-minecraft/src/main/java/net/guizhanss/guizhanlib/minecraft/localization/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides single-file localization service specialized for Minecraft.
+ */
+package net.guizhanss.guizhanlib.minecraft.localization;