+++ guizhanlib-minecraft/src/main/java/net/guizhanss/guizhanlib/minecraft/utils/InventoryUtil.java
@@ -19,12 +19,13 @@ import java.util.Objects;
 @SuppressWarnings("ConstantConditions")
 @UtilityClass
 public final class InventoryUtil {
+
     /**
      * This method push all {@link ItemStack} to {@link Player}'s inventory.
      * if some items cannot be pushed, they will be dropped at {@link Player}'s {@link Location}.
      *
-     * @param p          the {@link Player}
-     * @param itemStacks all the {@link ItemStack}s to be pushed
+     * @param p          the {@link Player}.
+     * @param itemStacks all the {@link ItemStack}s to be pushed.
      */
     @ParametersAreNonnullByDefault
     public static void push(Player p, ItemStack... itemStacks) {
@@ -37,9 +38,9 @@ public final class InventoryUtil {
      * This method push all {@link ItemStack} to {@link Player}'s inventory.
      * if some items cannot be pushed, they will be dropped at the specified {@link Location}.
      *
-     * @param p          the {@link Player}
-     * @param loc        the {@link Location} to drop items
-     * @param itemStacks all the {@link ItemStack}s to be pushed
+     * @param p          the {@link Player}.
+     * @param loc        the {@link Location} to drop items.
+     * @param itemStacks all the {@link ItemStack}s to be pushed.
      */
     @ParametersAreNonnullByDefault
     public static void push(Player p, Location loc, ItemStack... itemStacks) {